+++ mmysql/dif.go
@@ -12,7 +12,7 @@ func Dif(r *gin.Engine, relativePath string, dbName string, tableName string) {
 	DifPlus(r, relativePath, dbName, tableName, nil, nil)
 }
 
-func DifPlus(r *gin.Engine, relativePath string, dbName string, tableName string, funcParam func(c *gin.Context, param gin.H) (gin.H, int64, error), funcResult func(c *gin.Context, result []gin.H) ([]gin.H, int64, error)) {
+func DifPlus(r *gin.Engine, relativePath string, dbName string, tableName string, funcParam func(c *gin.Context, param gin.H) (gin.H, int, error), funcResult func(c *gin.Context, result []gin.H) ([]gin.H, int, error)) {
 	r.POST(relativePath, func(c *gin.Context) {
 		param, err := ParamToGinH(c)
 		if err != nil {
@@ -59,7 +59,7 @@ func DifPlus(r *gin.Engine, relativePath string, dbName string, tableName string
 	})
 }
 
-func MysqlDif(param gin.H, dbName string, tableName string) ([]gin.H, int64, error) {
+func MysqlDif(param gin.H, dbName string, tableName string) ([]gin.H, int, error) {
 	table := tableNameFromeParam(param, tableName)
 	if param["field"] == nil || len(param["field"].(string)) == 0 {
 		return nil, 10004, errors.New("field参数不能为空")
