+++ mmysql/list.go
@@ -13,7 +13,7 @@ func List(r *gin.Engine, relativePath string, dbName string, tableName string, s
 	ListPlus(r, relativePath, dbName, tableName, searchTargets, nil, nil)
 }
 
-func ListPlus(r *gin.Engine, relativePath string, dbName string, tableName string, searchTargets []string, funcParam func(c *gin.Context, param gin.H) (gin.H, int64, error), funcResult func(c *gin.Context, result []gin.H, count int64) ([]gin.H, int64, int64, error)) {
+func ListPlus(r *gin.Engine, relativePath string, dbName string, tableName string, searchTargets []string, funcParam func(c *gin.Context, param gin.H) (gin.H, int, error), funcResult func(c *gin.Context, result []gin.H, count int64) ([]gin.H, int64, int, error)) {
 	r.GET(relativePath, func(c *gin.Context) {
 		param := paramFromGet(c)
 		// 处理参数
@@ -103,7 +103,7 @@ func ListPlus(r *gin.Engine, relativePath string, dbName string, tableName strin
 	})
 }
 
-func MysqlList(param gin.H, dbName string, tableName string, searchTargets []string) ([]gin.H, int64, int64, error) {
+func MysqlList(param gin.H, dbName string, tableName string, searchTargets []string) ([]gin.H, int64, int, error) {
 	// 处理列表数据
 	table := tableNameFromeParam(param, tableName)
 	sortString := fmt.Sprintf("%s %s", "createTime", "DESC")
